+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Change history
 
 .. note::
 
-    A new ``JSONSCHEMA_USE_FORMAT_CHECKER`` setting has been introduced (default: ``True``) for newly created objecttypes.
+    A new ``JSONSCHEMA_USE_FORMAT_CHECKER`` setting has been introduced (default: ``False`` for existing objecttypes when upgrading to 4.2.0, default: ``True`` for newly created objecttypes after that).
     A new boolean field, ``strict_format_checker``, has also been added to the ObjectType model to control
     whether JSON Schema ``format`` validation is applied for that specific objecttype.
     When set to ``None``, the value of the ``JSONSCHEMA_USE_FORMAT_CHECKER`` setting is used.